Add http request as the last param for a method call (#17)
In some rpc call handler implementations it is useful to have the
original http request. One example would be to lookup the local address
or remote address of the caller so use this information in the response.

This change adds the http request as an optional last argument to the
rpc handler. The `HttpRequest` type is also updated with a `socket`
field containing `localAddress`, `localPort`, `remoteAddress`, and
`remotePort`. If the http server implementation is able to specify these
fields it can do so.